feat: l10n
added l10n + fixed tests
