+++ l10n.yaml
@@ -1,0 +1,3 @@
+arb-dir: lib/l10n
+template-arb-file: app_sk.arb
+output-localization-file: app_localizations.dart
