+++ test/app_test.dart
@@ -1,11 +1,19 @@
-import 'package:flutter_test/flutter_test.dart';
 import 'package:biely_kvet/main.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'utils.dart';
 
 void main() {
   testWidgets('App načíta hlavné menu', (tester) async {
     await tester.pumpWidget(const BielyKvetApp());
 
-    expect(find.text('Odpovede'), findsOneWidget);
-    expect(find.text('Pravidlá'), findsOneWidget);
+    expect(
+      find.text((await slovakAppLocalization).questionsTabTitle),
+      findsOneWidget,
+    );
+    expect(
+      find.text((await slovakAppLocalization).rulesTabTitle),
+      findsOneWidget,
+    );
   });
 }
