+++ test/utils.dart
@@ -1,0 +1,6 @@
+import 'dart:ui';
+
+import 'package:biely_kvet/l10n/app_localizations.dart';
+
+var slovakAppLocalization = AppLocalizations.delegate.load(Locale('sk'));
+
